requirements.txt
Flask==2.3.3
peewee==3.17.0
bcrypt==4.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
openpyxl==3.1.2